improve record self linking
when requiring the backing action to be named `:get` to be able to generate the self link, it is not imediately clear why the self link does not appear in the JSON:API response. Only after some digging and understanding that I would need to implement a specific action on all my resources for them to be able to link to themselfs, I found this link of Code.

I would like to suggest, to better utilize the default `:read` action many resources have.
The route selection filter for the self link should then only require the HTTP method to be `:get` which is in scope of this project and not enforce restrictions on the backing action name, which in my opinion should be out of scope of this project and up to the end user of the framework.