fix(OpenApi): Don't cause a compilation failure when open_api_spex is not present.
The `open_api_spex` package is set as an optional dependency, so when it's not present we need to not compile the OpenApi controller.  I considered having it return an error, but I think it's better to not include it in the routes at all if the package is missing and the server will return a 404.