improvement: properly represent fields in response
This requires two main changes: no property is required. This is because
you can always limit the fields using the `fields` parameter.

Additionally, we represent nillable types with `type | null`.
We now also include a comment in the description as to wether or not the field
is included by defaul.t

Finally, if a route has `default_fields`, we create a new in-line object type
for that route