Chore: move OpenAPI spec generation from controller to OpenAPI module
We strive to be able to generate the OpenAPI spec with Mix commands, and
not only by querying an HTTP route.

As a preliminar step for further developments, move the spec generation
to the OpenAPI module in order to have a more reusable `spec` function
instead of having it tied to the controller.

Notes:
- Even though the OpenAPI module is not strictly related to HTTP
connections, the spec function accepts a `conn` argument since it was
already needed and provided to the `modify_open_api` function, if the
user specifies it in the opts. The `conn` argument is nil-able since we
want to support spec generation from the CLI.
- The OpenAPI title is now configurable with the `:open_api_title` opt.
- The OpenAPI version is now configurable with the `:open_api_version`
opt.
- The OpenAPI server list was previously populated by getting the
endpoint from a live `conn`. Since we want to support spec generation
from the CLI and without access to an active connection, a new option
`:phoenix_endpoint` is supported to statically provide a reference to
the endpoint. Another option `:open_api_servers` is added if the user
instead wants to specify a custom list of URLs.

Signed-off-by: Davide Briani <davide@briani.dev>