Support generating OpenAPI spec files with Mix tasks (#200)
* Chore: move OpenAPI spec generation from controller to OpenAPI module

We strive to be able to generate the OpenAPI spec with Mix commands, and
not only by querying an HTTP route.

As a preliminar step for further developments, move the spec generation
to the OpenAPI module in order to have a more reusable `spec` function
instead of having it tied to the controller.

Notes:
- Even though the OpenAPI module is not strictly related to HTTP
connections, the spec function accepts a `conn` argument since it was
already needed and provided to the `modify_open_api` function, if the
user specifies it in the opts. The `conn` argument is nil-able since we
want to support spec generation from the CLI.
- The OpenAPI title is now configurable with the `:open_api_title` opt.
- The OpenAPI version is now configurable with the `:open_api_version`
opt.
- The OpenAPI server list was previously populated by getting the
endpoint from a live `conn`. Since we want to support spec generation
from the CLI and without access to an active connection, a new option
`:phoenix_endpoint` is supported to statically provide a reference to
the endpoint. Another option `:open_api_servers` is added if the user
instead wants to specify a custom list of URLs.

Signed-off-by: Davide Briani <davide@briani.dev>

* Feat: add `spec/0` on Router to support OpenAPI generation via CLI

If OpenApiSpex is present, the `spec/0` function is exposed on the
Router.
This allows a user who has already configured AshJsonApi

```elixir
defmodule MyAppWeb.AshJsonApi
  use AshJsonApi.Router, domains: [...], open_api: "/open_api"
end
```

to directly generate the OpenAPI spec by running one the Mix tasks from
OpenApiSpex:

```sh
mix openapi.spec.json --spec MyAppWeb.AshJsonApi
mix openapi.spec.yaml --spec MyAppWeb.AshJsonApi
```

Signed-off-by: Davide Briani <davide@briani.dev>

* Doc: describe how to customize OpenAPI spec and generate via CLI

Add some details about the available options to customize the values of
the OpenAPI spec.
Add a section to describe how the spec files can be generated with Mix
tasks.

Signed-off-by: Davide Briani <davide@briani.dev>

* Doc: remove reference to `title` opt for OpenApiSpex.Plug.SwaggerUI

The option is not supported by OpenApiSpex.Plug.SwaggerUI and has no
effect.
To configure the title of the OpenAPI spec, a specific option has been
added on AshJsonApi.Router.

Signed-off-by: Davide Briani <davide@briani.dev>

---------

Signed-off-by: Davide Briani <davide@briani.dev>