fixes two issues found: couldn't post through related; returned wrong id
This fixes two separate but couple issues I found while using the
package. It started by me not being able to create related resources
through the spec-approved path of /:parent_id/relationships/:relation
with a payload of `{"data":[{"id":"ID","type":"RELATION"}]}`. It appears
that while the code converts the keys to atoms, it was actually checking
for the string version of the key. Furthermore, any request containing
a `meta` tag failed.

While creating the tests that verified the actual fix for that issue,
another issue presented itself: when creating a related resource, the
server responded with the correct `type` but the wrong `id`. My first
thought was it was returning the _join table_ id of the related resource
but I actually now think it was returning the id of the parent. Either
way, I also fixed the issue and it now returns the correct `type`/`id`
combination.